refactored sources and mode computation
