+++ src/fdtdx/fdtd/update.py
@@ -3,7 +3,7 @@ import pytreeclass as tc
 
 from fdtdx.config import SimulationConfig
 from fdtdx.fdtd.container import ArrayContainer, ObjectContainer
-from fdtdx.fdtd.curl import curl_E, curl_H, interpolate_fields
+from fdtdx.core.physics.curl import curl_E, curl_H, interpolate_fields
 from fdtdx.fdtd.misc import add_boundary_interfaces, collect_boundary_interfaces
 from fdtdx.objects.boundaries.periodic import PeriodicBoundary
 from fdtdx.objects.detectors.detector import Detector
